% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aoe2_match.r
\name{aoe2_match}
\alias{aoe2_match}
\title{aoe2_match}
\usage{
aoe2_match(match_id, id_to_text = .pkgenv$id_to_text)
}
\arguments{
\item{match_id}{Either match_id or uuid of the match. match_id is the same as the Lobby code. Function takes either value and changes the api call accordingly.}

\item{id_to_text}{Translates columns with number code to humanreadable text. Like civ ids to civ names.}
}
\description{
This function gives you list of the match settings and match details.
}
\details{
More Information about aoe2.net api at https://aoe2.net/#api
}
\examples{
#Hera vs Liereyy Champions League
aoe2_match(match_id=47634872)
#Hera vs Liereyy Champions League
aoe2_match(match_id="39b80e7a-e4df-974e-90f0-cf4fe2be6561")
}
\keyword{aoe2_match}
