% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aoe2_match_rec_search.R
\name{aoe2_match_rec_search}
\alias{aoe2_match_rec_search}
\title{aoe2_match_rec_search}
\usage{
aoe2_match_rec_search(match_id)
}
\description{
This function checks if a game has recording available and returns an recording
}
\details{
More Information about aoe2.net api at https://aoe2.net/#api
}
\examples{
aoe2_match_rec_search()

}
\keyword{aoe2_match_rec_search}
