% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aoe2_matches.r
\name{aoe2_matches}
\alias{aoe2_matches}
\title{aoe2_matches}
\usage{
aoe2_matches(date, count = 10, game = "aoe2de", id_to_text = id_to_text)
}
\arguments{
\item{date}{Date either as unix epoch or in this format "2013-09-16 2:13:46 EST"}

\item{count}{Number of games to download.}

\item{game}{Default is aoe2de since the api is mostly used for that.}

\item{id_to_text}{Translates columns with number code to humanreadable text. Like civ ids to civ names.}
}
\description{
This function gives you data.frame of the player rating if the player has been active in the last 28 days.
}
\details{
More Information about aoe2.net api at https://aoe2.net/#api
}
\examples{
aoe2_matches(date="2020-11-09 12:49:39 UTC", count=5)
aoe2_matches(date="1596775000", count=5) # recommended way
}
\keyword{matches}
