% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aoe2_number_of_players_online.r
\name{aoe2_number_of_players_online}
\alias{aoe2_number_of_players_online}
\title{aoe2_number_of_players_online}
\usage{
aoe2_number_of_players_online(game = "aoe2de")
}
\arguments{
\item{game}{Default is aoe2de since the api is mostly used for that.}
}
\description{
This function gives you a list of player activity of the last 28 days for every 5min.
}
\details{
More Information about aoe2.net api at https://aoe2.net/#api
}
\examples{
aoe2_number_of_players_online()
}
\keyword{aoe2_number_of_players_online}
