% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aoe2_online.R
\name{aoe2_online}
\alias{aoe2_online}
\title{aoe2_online}
\usage{
aoe2_online()
}
\description{
This function checks if aoe2.net is available.
}
\details{
More Information about aoe2.net api at https://aoe2.net/#api
}
\examples{
aoe2_online()

}
\keyword{aoe2_online}
