% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aoe2_player_match_history.r
\name{aoe2_player_match_history}
\alias{aoe2_player_match_history}
\title{aoe2_player_match_history}
\usage{
aoe2_player_match_history(
  player_id,
  start = 0,
  count = 10,
  game = "aoe2de",
  id_to_text = .pkgenv$id_to_text
)
}
\arguments{
\item{player_id}{Either profile_id or 64 bit steam id. Function takes either value and changes the api call accordingly}

\item{start}{Match at which to start counting. Default = '0'}

\item{count}{Number of games to download. Max 1000}

\item{game}{Default is aoe2de since the api is mostly used for that.}

\item{id_to_text}{Translates columns with number code to humanreadable text. Like civ ids to civ names.}
}
\description{
This function gives you data.frame of the player rating if the player has been active in the last 28 days.
}
\details{
More Information about aoe2.net api at https://aoe2.net/#api
}
\examples{
#Rating Match History of GL. TheViper with profile_id
player_match_history(player_id=196240, count=5)
#Rating Match History of GL. TheViper with steam_id
player_match_history(player_id=76561197984749679, count=5)
}
\keyword{player_rating}
