% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aoe2_player_profile_all.r
\name{aoe2_player_profile_all}
\alias{aoe2_player_profile_all}
\title{aoe2_player_profile_all}
\usage{
aoe2_player_profile_all(player_id, game = "aoe2de")
}
\arguments{
\item{player_id}{Either profile_id or 64 bit steam id. Function takes either value and changes the api call accordingly}

\item{game}{Default is aoe2de since the api is mostly used for that.}
}
\description{
This function gives you all available ranked player data. (The goal is it to have a function which creates a player object on which it would be easier to perfom regular summary funcitons)
}
\details{
More Information about aoe2.net api at https://aoe2.net/#api
}
\examples{
#Profile of GL. TheViper with profile_id
aoe2_player_profile(player_id=196240)
#Profile of GL. TheViper with steam_id
aoe2_player_profile(player_id=76561197984749679)
}
\keyword{aoe2_player_profile}
