% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/id_to_text.r
\name{id_to_text}
\alias{id_to_text}
\title{id_to_text}
\usage{
id_to_text(data = data, id_to_text = .pkgenv$id_to_text)
}
\arguments{
\item{data}{aoe2.net output}

\item{id_to_text}{Translates columns with number code to humanreadable text. Like civ ids to civ names.}
}
\description{
This function translates the values to human readable. eg. map 1= arabia.
}
\details{
More Information about aoe2.net api at https://aoe2.net/#api
}
\keyword{id_to_text}
