% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lobbies.r
\name{lobbies}
\alias{lobbies}
\title{lobbies}
\usage{
lobbies(game = "aoe2de")
}
\arguments{
\item{game}{Default is aoe2de since the api is mostly used for that.}
}
\description{
This function gives you list of open Lobbies.
}
\details{
More Information about aoe2.net api at https://aoe2.net/#api
}
\examples{
#List of open Lobbies
lobbies()
}
\keyword{lobbies}
