% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.r
\docType{data}
\name{lobby_translation}
\alias{lobby_translation}
\title{All importent keys to translate aoe2api stuff to humanreadable code.}
\format{
A list containing stuff
\describe{
Used internaly to translate Ids to names of maps, civs, play_modes etc.
}
}
\source{
\url{https://aoe2.net/api/strings?game=aoe2de&language=en}
}
\usage{
lobby_translation
}
\description{
All importent keys to translate aoe2api stuff to humanreadable code.
}
\keyword{datasets}
