% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/match.r
\name{match}
\alias{match}
\title{match}
\usage{
match(match_id)
}
\arguments{
\item{match_id}{Either match_id or uuid of the match. match_id is the same as the Lobby code. Function takes either value and changes the api call accordingly.}
}
\description{
This function gives you list of the match settings and match details.
}
\details{
More Information about aoe2.net api at https://aoe2.net/#api
}
\examples{
#Hera vs Liereyy Champions League
match(match_id=47634872)
#Hera vs Liereyy Champions League
match(match_id="39b80e7a-e4df-974e-90f0-cf4fe2be6561")
}
\keyword{match}
