% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.r
\docType{data}
\name{match_translation}
\alias{match_translation}
\title{All civs in aoe2de.}
\format{
A list containing stuff
\describe{
Used internaly to translate Ids to names of maps, civs, play_modes etc.
}
}
\source{
\url{https://aoe2.net/api/strings?game=aoe2de&language=en}
}
\usage{
match_translation
}
\description{
All civs in aoe2de.
}
\keyword{datasets}
