% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/player_last_match.r
\name{player_last_match}
\alias{player_last_match}
\title{player_last_match}
\usage{
player_last_match(player_id, game = "aoe2de")
}
\arguments{
\item{player_id}{Either profile_id or 64 bit steam id. Function takes either value and changes the api call accordingly}

\item{game}{Default is aoe2de since the api is mostly used for that.}
}
\description{
This function gives you data.frame of the player rating if the player has been active in the last 28 days.
}
\details{
More Information about aoe2.net api at https://aoe2.net/#api
}
\examples{
#Last Match of GL. TheViper with profile_id
player_last_match(player_id=196240)
#Last Match of GL. TheViper with steam_id
player_last_match(player_id=76561197984749679)
}
\keyword{player_rating}
