% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/player_match_history_all.r
\name{player_match_history_all}
\alias{player_match_history_all}
\title{player_match_history_all}
\usage{
player_match_history_all(player_id, game = "aoe2de", id_to_text = TRUE)
}
\arguments{
\item{player_id}{Either profile_id or 64 bit steam id. Function takes either value and changes the api call accordingly}

\item{game}{Default is aoe2de since the api is mostly used for that.}

\item{id_to_text}{Translates columns with number code to humanreadable text. Like civ ids to civ names.}
}
\description{
This function gives you data.frame of the player rating if the player has been active in the last 28 days.
}
\details{
More Information about aoe2.net api at https://aoe2.net/#api
}
\examples{
#Rating Match History of GL. TheViper with profile_id
player_match_history_all(player_id=196240)
#Rating Match History of GL. TheViper with steam_id
player_match_history_all(player_id="76561197984749679")
}
\keyword{player_match_history_all}
