% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/player_rating_history.r
\name{player_rating_history}
\alias{player_rating_history}
\title{player_rating_history}
\usage{
player_rating_history(
  player_id,
  start = 0,
  count = 10,
  leaderboard_id = 3,
  game = "aoe2de"
)
}
\arguments{
\item{player_id}{Either profile_id or 64 bit steam id. Function takes either value and changes the api call accordingly.}

\item{start}{Match at which to start counting. Default = '0'}

\item{count}{Number of games to download. Max 1000.}

\item{leaderboard_id}{Default is 3 for 1v1 Random Map, 4 is Team Random Map.}

\item{game}{Default is aoe2de since the api is mostly used for that.}
}
\description{
This function gives you data.frame of the player rating if the player has been active in the last 28 days.
}
\details{
More Information about aoe2.net api at https://aoe2.net/#api
}
\examples{
#Rating History of GL. TheViper with profile_id
player_rating_history(player_id=196240, count=5)
#Rating History of GL. TheViper with steam_id
player_rating_history(player_id=76561197984749679, count=5)
}
\keyword{player_rating_history}
