% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/strings.r
\name{strings}
\alias{strings}
\title{strings}
\usage{
strings(language = "en")
}
\arguments{
\item{language}{Default is en. Other values are en, de, el, es, es-MX, fr, hi, it, ja, ko, ms, nl, pt, ru, tr, vi, zh, zh-TW}
}
\description{
This function gives you a list with the human readable values to the data downloaded by this package.
Eg civs 0='Aztecs', 1='Berbers', ...
Eg age 0='standard', 1='Dark Age'
}
\details{
More Information about aoe2.net api at https://aoe2.net/#api
}
\examples{
#
strings(language="de")
strings(language="lt") #Lithuanian is not supported
}
\keyword{strings}
